Add routing.go to separate routes from router
